requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
rdkit
networkx
matplotlib
Pillow
openpyxl
pydot 